requirements.txt
fastapi
uvicorn
pandas
numpy
scikit-learn
imbalanced-learn
xgboost
joblib
groq
python-dotenv
